research/semantic_fluency_analysis/requirements.txt
# Core data science libraries
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0

# Machine learning and clustering
scikit-learn>=1.0.0

# Natural language processing
spacy>=3.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Statistical analysis
statsmodels>=0.13.0

# Configuration and utilities
pyyaml>=6.0

# Optional: For advanced statistical analysis
# pingouin>=0.5.0  # Uncomment if you want additional statistical tests
